ADC changed to PA4. I2C changed to PA12(SCL) & PA15(SDA), no more hardware I2C support. PPM changed to 8 channel output(PB0,1,4,5,6,7,8,9) and 6 channel input(PB15,10,11,12,13,14,15), pilot changed accordingly. station: time changed to 64bit.

+++ station.cpp
@@ -160,7 +160,7 @@ int main(void)
 	int64_t last_packet_time = 0;
 	int packet_speed = 0;
 	int packet_speed_counter = 0;
-	int packet_speed_time = getus();
+	int64_t packet_speed_time = getus();
 	while(1)
 	{
 		int result = NRF_Rx_Dat((u8*)&recv);
